Graphic : Add type cast

+++ xxGraphicD3D10.cpp
@@ -313,7 +313,7 @@ uint64_t xxCreateRenderPassD3D10(uint64_t device, bool clearColor, bool clearDep
     if (clearStencil)
         d3dRenderPass.clearDepthStencil |= D3D10_CLEAR_STENCIL;
 
-    return d3dRenderPass.value;
+    return static_cast<uint64_t>(d3dRenderPass.value);
 }
 //------------------------------------------------------------------------------
 void xxDestroyRenderPassD3D10(uint64_t renderPass)
